[NCL-8048] Add authentication token when required
This commit adds supports for the `authTokenRequired` boolean variable
to the execute-rest long process. This variable (default: false) injects
a fresh OIDC access token to the HTTP Headers before making a request.

This is needed because access tokens have a short span of life and needs
to be renewed to be valid.

We cannot rely on the access token to be passed from the
`requestHeaders` because with retries, the access token in the
`requestHeaders` might already expire.